requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
celery==5.5.3
redis==6.2.0
psycopg2-binary==2.9.10
django-filter==25.1
drf-yasg==1.21.10
python-decouple==3.8